+++ lib/hammer/redis/token_bucket.ex
@@ -65,7 +65,7 @@ defmodule Hammer.Redis.TokenBucket do
       use Hammer, backend: Hammer.Redis, algorithm: :token_bucket
       end
 
-      MyApp.RateLimit.start_link(clean_period: :timer.minutes(1))
+      MyApp.RateLimit.start_link([])
 
       # Allow 10 tokens per second with max capacity of 100
       MyApp.RateLimit.hit("user_123", 10, 100, 1)