+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## 7.1.0 - 2025-08-05
+
+### Added
+
+- Implement sliding window algorithm with support for increment > 1
+- Add inc/6 and set/6 functions for sliding window algorithm
+
 ## 7.0.2 - 2025-02-10
 
-### Changed
+### Fixed
 
 - Fix incorrect timeout typespec